+++ common/db/src/main/java/is/codion/common/db/pool/AbstractConnectionPoolWrapper.java
@@ -45,6 +45,8 @@ public abstract class AbstractConnectionPoolWrapper<T> implements ConnectionPool
 	private static final String GET_CONNECTION = "getConnection";
 	private static final String CLOSE = "close";
 
+	private static final boolean VALIDATE = VALIDATE_CONNECTIONS_ON_CHECKOUT.getOrThrow();
+
 	/**
 	 * The actual connection pool object
 	 */
@@ -81,7 +83,7 @@ public abstract class AbstractConnectionPoolWrapper<T> implements ConnectionPool
 		try {
 			counter.incrementRequestCounter();
 
-			return fetchConnection();
+			return validate(fetchConnection());
 		}
 		catch (SQLException e) {
 			counter.incrementFailedRequestCounter();
@@ -122,6 +124,18 @@ public abstract class AbstractConnectionPoolWrapper<T> implements ConnectionPool
 	@Override
 	public final ConnectionPoolStatistics statistics(long since) {
 		return counter.collectStatistics(since);
+	}
+
+	private Connection validate(Connection connection) throws SQLException {
+		if (VALIDATE && !connectionFactory.connectionValid(connection)) {
+			try {
+				connection.close();
+			}
+			catch (SQLException ignored) {}
+			throw new SQLException("Connection retrieved from pool is invalid");
+		}
+
+		return connection;
 	}
 
 	/**