+++ changelog.md
@@ -12,6 +12,7 @@ Codion Change Log
 - AbstractConnectionPoolWrapper.closeStatisticsCollection() added.
 - SLF4J dependency added, DefaultDatabaseConnection logging improved.
 - ConnectionPoolWrapper.DEFAULT_CHECK_OUT_TIMEOUT added.
+- ConnectionPoolWrapper.VALIDATE_CONNECTIONS_ON_CHECKOUT added.
 ### is.codion.dbms.h2
 - H2Database, security validation for the H2 database script injection.
 ### is.codion.swing.common.ui